V2XSW-72: Convert public key from hsm to v2xse api format
Public keys returned from the hsm need to be converted to v2xse api
format.  The hsm format has the x coordinate directly followed by the
y coordinate.  The v2xse api format allocates enough space for a 384
bit x and y coordinate in all cases, so in the case of a 256 bit key
there are 128 unused bits between the end of the x coordinate and the
start of the y coordinate.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>