V2XSW-88: Modify memory clear to be easier to understand
Static analysis tool finding: Out-of-bounds access.  This was intentional,
but needed the reader to know the structure layout in order to understand
that it was starting the clear after the last valid entry.  The code has
been modified to clear the static analysis warning, and also makes it
easier to understand that the clear starts after the last entry of the
hsm structure.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>