V2XSW-88: Remove existence checks when handling nvm entries
Static code analysis finding: "Time of check time of use".  When handling
nvm entries, there were tests for existence before many accesses.  The
analysis tool complained because there is a window of time between the
check and the access where the fs state could change, rendering the check
result invalid.  As reading items that do not exist give an error anyway,
these checks were redundant for reads, so have been removed.  When writing
variables, there were no checks as the standard operation is overwrite.
When creating directories, where previously the code checked if the
directory existed and if not created it, this was changed to always trying
to create the directory, and ignoring the error if errno indicated the
error was due to the directory already existing.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>