V2XSW-72: Clear y bits 32 - 47 when converting pubkey to v2xSe format
When converting a public key from hsm API to v2xSe API for a 256 bit
curve, bits 32 - 47 of the x and y components are unused.  The unused
x bits were already cleared, this change also clears the unused y bits.
This is useful for testing as the whole key structure will have known
values and can be compared to reference values without needing to work
out which values are relevant or not.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>