V2XSW-74: Do not use pointer to ECIES output size on input
When calling v2xSe API ECIES encrypt or decrypt functions, the values
pointed to by the output size pointer should not be used.  As the maximum
size of the output of these functions is defined in the API, overflow is
avoided by the caller making sure the buffer is at least as big as the
maximum size.  The commit removes the use of these values on input,
instead the values required by the hsm are used.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>