V2XSW-71: Match SXF1800 error codes when device not activated
The v2xSe Crypto API indicates that the device must be activated for many
commands, and indicates that the error code V2XSE_DEVICE_NOT_CONNECTED
should be used if not. This is tested in the SXF1800 test suite when the
device is in INIT state.  A different test performs the same operation
when the device is in CONNECTED state, and expects different results,
ie error code V2XSE_FAILURE and status code V2XSE_INACTIVE_CHANNEL.  It
seems logical to not give error V2XSE_DEVICE_NOT_CONNECTED when in
CONNECTED state, so we will use the same error codes for this case
in the adaptation layer.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>