V2XSW-19: Set default status for pHsmStatusCode
Set *pHsmStatusCode to V2XSE_UNDEFINED_ERROR by default.  This removes
the need to explicitly set this value for non-specific errors, and
also catches a couple of cases where no code was set.  The setting
is done in the macro that verifies that the pointer is not NULL, which
has been renamed VERIFY_STATUS_PTR_AND_SET_DEFAULT to reflect its new
operation.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>