V2XSW-88: Remove use of system function
Static code analysis considers the use of the system function a security
hole.  Rework the deletion of nvm contents to no longer use a system
call.  The old implementation executed "rm -rf *" in the nvm directory
for the current applet.  The new implementation removes all regular files
from the cureent applet's nvm directory as well as from the four
subdirectories corresponding to the arrays for ba/rt keyHandle/curveId.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>