V2XSW-269: Replace use of 'int' with 'int32_t'
Full misra compliance requires that type 'int' not be used, but a
typedef that specifies the size of the underlying storage be used.
As the base type used for the v2xSeApi is int32_t, replacing all
instances of int with int32_t in the code makes sense.  New code
added in the last few months have generally used int32_t, this commit
makes this consistent in the whole code.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>