V2XSW-131: Avoid modifying input public key for v2xSe_encryptUsingEcies
The previous implementation of v2xSe_encryptUsingEcies modified the public
key passed to it (changing from v2xSe to HSM format).
This can lead to hard-to-find issues where using the same copy of the
public key again gives unexpected results.
To avoid this, v2xSe_encryptUsingEcies has been changed so that the
input public key is not modified.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>