V2XSW-208: Workaround for butterfly update issue in HSM PRC2 code
When the HSM PRC2 code performs a butterfly derivation to update a key,
it updates the base key instead of the derived key.  Creating a new key
using butterfly derivation works as expected.  To workaround the update
issue, always delete the existing key if it exists, so an update will
never occur.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>