V2XSW-480 Implement SM2 Key exchange
These API functions use the generic hsm_ey_exchange() API from SECO libs
for the  SM2 key generation (as specified in GB/T 32918) use case.

For this use case, the HSM only supports the role of the responder. The
user of this library thus must be the initiator of the key exchange operation.

Both entities have to provide two keys: a static key and an ephemeral
key, public keys as well as Z values are exchanged during the process.

The resulting shred key can be of type SM2 or SM4 but the key exchange
data type must be SM2 (assymmetric keys used for exchange).

Signed-off-by: Stephane Viau <stephane.viau@nxp.com>